refactor(daemon): extract GitHub hook dispatch and review orchestration
Move the GitHub hook-dispatch and formal-review cluster out of the Daemon
class into src/github/review-orchestrator.ts, following the established
class + narrow host + thin same-name delegate template (evaluation/daemon-hooks.ts,
webchat/, commands/handlers.ts).

Relocated by name: dispatchRelayHook, onHookFire,
completeGithubThreadWorktreeCleanup, githubFormalReviewEnabled,
ensureGithubPullRevision, githubWorkspaceMatches, sessionInitiatorLabel,
prepareGithubReviewWorkspace, prepareGithubTurn, persistGithubReviewEffect,
reconcileGithubReviewAttempt, submitGithubReview, replyGithubReviewThreads,
makeGithubReply, githubCommentAttribution. The githubReviewClient field moves
with them; the Daemon keeps a getter onto it.

The activeGithubTurnMeta / activeGithubReplyBatchMeta maps stay on the Daemon
(dispatchOne finalizes turns from them) and are read through the host, as does
the generic hook state/report plumbing (persistInbox, persistHookState,
emitHookCompletion). makeGithubReply is reached back through the host so a
per-turn override on the Daemon is still honored. No behavior change.
